* Fixed a problem with tinting effect * All layers are modified to call their base * All layers can be created without coordinates * Old effect definitions are removed from GGEMain * Setup function added to GGEMain allowing to adjust screen   properties after init * Some layers can be initialized by supplying a rect * An unused data in set texture is removed, might cause problems * BRG is added to graphics.h to ensure portability * Leak caused by unregistering keyboard events is fixed * Adding a layer to another parent detaches the child from the former parent * OS depended sleep function added * const correctness for DisplayMessage is fixedmode * Shadow class has offset parameter as point, instead of two separate vars * Accessing Width/Height of images is blocked, getWidth and getHeight should be used * Font object to define display properties of a font * Font renderer interface (which BitmapFontResource is derived from) * Font theme resource that specifies different renderers for different styles * Data resource can return Size2D and Margins data * Folders sort their base items correctly * Collection is modernized   - Memory leak by using collection iterators is fixed   - Growth parameter of Collection is now a template parameter   - Collection copy constructor   - Search function * EventChain is expanded (+480lines on top of 250)   - More C++ friendly   - Allows different handler function signatures * LinkedList is modernized and expanded   - Iteration type support to be used with foreach   - Iterator object do not leak memory anymore   - Both growth and sortkey are now template params   - More C++ friendly * ManagedBuffer uses exceptions properly * Button's SetBluePrint function is fixed * Frame got itself a SetBluePrint function * Content size property is added to Frame * Linear organizer knows widths of every line * WidgetRegistry auto created object is renamed to WR * Some of sprintf functions are replaced by sprintf_s * Link Data * Font Data * File system helper added with a function to check if a file exist

+++ Effects/LayerMover.h
@@ -25,7 +25,7 @@ namespace geffects {
 		////Initializes the effect
 		LayerMover(LayerBase *Target) :
 			speed(0,0),
-			current(Target->X, Target->Y),
+			current((float)Target->X, (float)Target->Y),
 			Target(Target), FinishedEvent("Finished", this)
 		{
 			AnimatorBase::FinishedEvent=FinishedEvent;
@@ -33,7 +33,7 @@ namespace geffects {
 		////Initializes the effect
 		LayerMover(LayerBase &Target) :
 			speed(0,0),
-			current(Target.X, Target.Y),
+			current((float)Target.X, (float)Target.Y),
 			Target(&Target), FinishedEvent("Finished", this)
 		{
 			AnimatorBase::FinishedEvent=FinishedEvent;
